+++ source/data.js
@@ -10,8 +10,6 @@ import {
 
 export const pwd = cwd()
 
-export const typesDir = 'compiled-types'
-
 // fill this with a map of dependency package names to versions that are busted
 // so that if they are necessary, a previous version is used instead
 export const bustedVersions = {}